Created a web component registry and exports components as modules in order to facilitate their customization
