+++ src/components/chat_content.js
@@ -3,9 +3,9 @@ import xss from "xss/dist/xss";
 import { CustomElement } from './element.js';
 import { html } from 'lit-element';
 import { unsafeHTML } from 'lit-html/directives/unsafe-html.js';
+import { api } from "@converse/headless/converse-core";
 
-
-class ChatContent extends CustomElement {
+export default class ChatContent extends CustomElement {
 
     static get properties () {
         return {
@@ -27,4 +27,4 @@ class ChatContent extends CustomElement {
     }
 }
 
-customElements.define('converse-chat-content', ChatContent);
+api.elements.define('converse-chat-content', ChatContent);
